requirements.txt
einops
essentia
g2pk
jamo
librosa
matplotlib
miditoolkit
praat-parselmouth==0.3.3
pyloudnorm
pyworld
pyyaml
resemblyzer
scikit-image
scipy
tensoarboard
textgrid
torchaudio
tqdm
webrtcvad